fixed nextPage being counted as a post

+++ admin/js/main.js
@@ -80,23 +80,23 @@ $(function(){
                 } else {
                     $.jGrowl(data)
                 }
-            },
+            }
         });
     };
 
-    grabPosts = function(pageNum, numOfItems) {
+    grabPosts = function() {
         var posts = $("<table style='width:100%'><tr><td><b>Heading</b></td><td><b>Created On</b></td><td><b>Action</b></td></tr></table>");
         $("#post-list").html(posts);
         $.ajax({
             type: 'POST',
             url: '../json.php',
             data: {
-                page: pageNum,
-                limit: numOfItems
+                limit: 0
             },
             dataType: 'json',
             success: function (data) {
                 for (i in data) {
+                    if(i == "nextPage") continue;
                     posts.append("<tr id=\"listedpostid" + data[i].id + "\"><td class='heading'>" + data[i].title + "</td> <td>" + data[i].time + "</td><td><a href='javascript: delete(" + data[i].id + ");'>Delete</a> | <a href='javascript: edit(" + data[i].id + ");'>Edit</a></td></tr>");
                     posts.find("#listedpostid" + data[i].id).data("post", data[i]);
                 }
